Reduce size of Poseidon machine (#2084)
After analyzing what is taking such a long time when evaluating fixed
column (#2083), I noticed that the Poseidon machine uses the default
degree range, independent of the `--max-degree-log` degree log.

I set it to `LARGE_SUBMACHINES_MAX_DEGREE` instead, which means that it
is 4x larger than the main machine. Given a block size of 31, that means
that it has enough rows if at most 1 / 8 (=12.5%) of the CPU
instructions are calls to Poseidon, which should be enough?

For `--max-degree-log 18`, this reduces the time to evaluate constant
columns from 109s to 33s.