Linker bus support (#2073)
Refactor of the linker to enable creating statements both in the source
and in the target of a link. Instead of creating a vector of statements,
we create a (name=>namespace) map.

The linker mode is exposed in the CLI, and currently defaults to the
native lookup/permutation statements.

Bus linking can be activated the following way:

```console
cargo run pil test_data/asm/vm_to_block_unique_interface.asm --force --linker-mode bus
```

TODO:
- [x] absolute names