Lower constraint degree for 256-Bit arithmetic machine (#2110)
This PR reduces the constraint degree of the 256-Bit arithmetic machines
from 4 to 3, making it possible to create Plonky3 proofs.

As explained in the comments, the current way to do it is not optimal,
because it adds roughly 256 witness columns. There would be an
alternative that only adds 5, but witgen doesn't currently work for
that.

This is a comparison for `test_data/std/arith256_memory_large_test.asm`:
| Metric                        | `main` | `fix-arith` |
|-------------------------------|--------|-------------|
| Fixed columns (arith)         | 32     | 32          |
| Witness columns (arith)       | 195    | 443         |
| Number of constraints (arith) | 234    | 482         |
| Max constraint degree         | 4      | 3           |
| Time witgen (BN254)           | 12.18s | 22.17s      |
| Time witgen (GL)              | 5.24s  | 7.26s       |
| Halo2-Composite proof time    | 0.91s  | 1.65s       |
| Plonky3 proof time            | N/A    | 0.38s       |