Add `PhantomBusInteraction` (#2183)
This PR adds a `Constr:: PhantomBusInteraction` variant. For now, it is
ignored - if users want to use a bus, they need to express this in terms
of phantom lookups / permutations as before this PR.

I added a few `TODO(bus_interaction)` and opened #2184 to track support
for phantom bus interactions.

One use-case this could have before though is to trigger a
"hand-written" witness generation for the bus, as discussed in the chat.