Call JIT (#2242)
This PR puts together the pieces to run compile-time witgen for block
machines. There are still many cases where it doesn't work yet, in which
case it falls back to run-time solving. These cases should be fixed in
future PRs.

It also fixes two bugs:
- When multiplying two affine expression, the case where one of them is
zero is now handled properly.
- `WitgenInference` now handles intermediate columns.

Note that this PR could slow down witgen by attempting to compile code
once per incoming connection and input / output combination, in block
machines. I think this should be negligible though and it gives us that
much of the new pipeline is already running in the tests and elsewhere.

# Benchmark results

I tested the code with different opt levels on a benchmark that computes
ca. $2^{16}$ Poseidon hashes.

## Baseline

```
 == Witgen profile (393220 events)
   93.0% (   30.8s): Secondary machine 0: main_poseidon (BlockMachine)
    4.1% (    1.4s): witgen (outer code)
    2.3% ( 750.8ms): Main machine (Dynamic)
    0.6% ( 204.4ms): FixedLookup
    0.0% (   3.2µs): range constraint multiplicity witgen
  ---------------------------
    ==> Total: 33.109672458s
```

## JIT (opt level 1)

```
 == Witgen profile (393222 events)
   52.3% (    7.7s): JIT-compilation
   32.0% (    4.7s): Secondary machine 0: main_poseidon (BlockMachine)
    9.2% (    1.3s): witgen (outer code)
    5.1% ( 748.3ms): Main machine (Dynamic)
    1.4% ( 213.5ms): FixedLookup
    0.0% ( 417.0ns): range constraint multiplicity witgen
  ---------------------------
    ==> Total: 14.729149333s
```

## JIT (opt level 3)

```
== Witgen profile (393222 events)
   94.6% (  107.9s): JIT-compilation
    3.4% (    3.9s): Secondary machine 0: main_poseidon (BlockMachine)
    1.1% (    1.3s): witgen (outer code)
    0.7% ( 746.5ms): Main machine (Dynamic)
    0.2% ( 204.1ms): FixedLookup
    0.0% ( 542.0ns): range constraint multiplicity witgen
  ---------------------------
    ==> Total: 114.036571291s
```