JIT: ignore bus interactions & reduce block machine row range (#2322)
I noticed that JITing never succeeds if we run with `--linker-mode bus`.
This PR fixes it, as can be verified by running this with debug log:
```
$ RUST_LOG=debug cargo run pil test_data/std/binary_large_test.asm -o output -f --linker-mode bus
...
Secondary machine 0: main_binary (BlockMachine): 18 / 18 blocks computed via JIT.
72 of 128 rows are used in machine 'Secondary machine 0: main_binary (BlockMachine)'.

 == Witgen profile (484 events)
   67.2% (    5.6s): JIT-compilation
   19.1% (    1.6s): Secondary machine 0: main_binary (BlockMachine)
   11.2% ( 944.5ms): FixedLookup
    1.5% ( 124.4ms): multiplicity witgen
    0.8% (  69.3ms): witgen (outer code)
    0.1% (  10.4ms): Main machine (Dynamic)
  ---------------------------
    ==> Total: 8.397369458s
```

Doing so, I came across an issue cause by the block machine processor
processing rows `-1..(self.block_size + 1)`: It can actually write to
the first row of the next block, in which case we
[panic](https://github.com/powdr-labs/powdr/blob/1ee15050a987beacd0eea36b7384bade48c51936/executor/src/witgen/jit/function_cache.rs#L124),
because we don't have that row yet. It would also be pretty annoying to
need to support that and not needed in practice. But it turns out, this
row range was actually more generous that what we use in the run-time
solver, see [this
line](https://github.com/powdr-labs/powdr/blob/1ee15050a987beacd0eea36b7384bade48c51936/executor/src/witgen/sequence_iterator.rs#L41)!
So I'm reducing it in this PR.