Fix `try_into_i32` (#2348)
`try_into_i32` is wrong for small fields, it always returns a positive
number.
I checked where it's used, and I think it makes more sense to have
`to_signed_integer`, so I changed that.
The only place it's used now is in the mock prover: negative values are
interpreted as receives, positive as sends. This would break if we
sent/received more than p/2 times, which seems fine for now.

Update: now also used in the riscv executor