Make latch semantic clearer (#2358)
Depends on #2360

Changes the semantics of the `PhantomBusInteraction.latch` field
slightly, to make my life easier with #2351. Basically, we want it to be
exactly what the RHS selector would have been in a lookup / permutation:
A fixed column latch for a lookup, and equal to the multiplicity for the
permutation. This way, witgen can treat it just like it treats selectors
now.

Eventually, I think we should remove the latch again. The permutation
latch is equal to the multiplicity anyway, and the lookup latch should
be detectable by looking for a constraint like `(1 - latch) *
multiplicity = 0;`.