Fix fixed lookup that contains constants (#2371)
This is something I came across in #2361: The `FixedLookupMachine` does
not recognize lookups of the form `[...] in [col1, col2, 3, col4]`. The
reason that this was not an issue before is that the PIL optimizer
simplifies such lookups. But it does not if they are represented as bus
interactions.

With this PR, the `FixedLookupMachine` looks for either fixed columns or
compile-time constants on the RHS.