Bootloader fix (#2364)
Currently, our bootloader setup doesn't take into acount the cost of
running the bootloader while setting up the inputs.
It may happen, in memory heavy workloads, that too many pages get paged
in by the bootloader and there is no rows left in the chunk for running
any computation.
This PR takes into account the cost of the bootloader when building its
inputs.