Witgen: Use bus natively (#2361)
Depends on #2371

This PR introduces a new `Identity`, which is similar to
`powdr_ast::analyzed::Identity`, but it only has "bus send" and "bus
receive" variants, instead of (phantom) lookup / permutation / bus
interaction. Any lookup or permutation gets translated to an equivalent
pair of bus interactions. This abstracts how the machine connection is
implemented exactly.

Right now, we assume that each bus send and receive has a compile-time
interaction ID, by which they can be statically matched. This is always
the case if the interaction originated from a lookup or permutation, so
this is fully backward-compatible.

But what this already gives us is many-to-one relationships: We can have
several sends with the same interaction ID, that map to the same
receive. As a result, we can drastically remove the number of receives,
as Thibaut is doing in #2359.

As a next step, we can also make the send interaction ID dynamic, which
allows for a reduction of sends. For that, we need to relax the
assumption that the interaction ID is known at compile time and match
while solving the constraint system.

### Benchmark results
```
executor-benchmark/keccak
                        time:   [11.928 s 12.248 s 12.611 s]
                        change: [-5.4383% -2.0982% +1.2884%] (p = 0.26 > 0.05)
                        No change in performance detected.
```

---------

Co-authored-by: chriseth <chris@ethereum.org>