Fix debug formatter. (#2398)
The debug formatter prints, for each identity, three lines of
unsimplified expressions (i.e. `x * 0` explicitly printed):
The first line contains just the identity, the second known values and
the third known range constraints.
Then it prints three more lines, with `x * 0` simplified as `0`. Again,
the first line with the simplified expression, the second with known
values, the third with known range constraints, but of course just for
the remaining variables.

For the second triple, it did not print the separators in the first line
(it did it properly for binary operators, though). Furthermore, it added
spacing to align the second triple with the first, even though the idea
is that this is not needed because the un-simplified version is often
not readable since it spans multpile lines.