Reverse fingerprint order (#2405)
This PR changes the fingerprint of a tuple $(x_1, ..., x_n)$:
from
$\sum_{i=1}^n \alpha^{(n - i)} x_i$
to
$\sum_{i=1}^n \alpha^{(i - 1)} x_i$

The difference is that previously, you could pad any number of zeros to
the left and get the same fingerprint. Now, you can pad any number of
zeros to the right, which I find more intuitive.

This also fixes a potential soundness bug: In practice, we always use
`fingerprint_with_id`, which prepends the bus ID to the payload, e.g.
`(<BUS_ID_FOO>, a, b, c)` & `(<BUS_ID_BAR>, a, b)`. But
`fingerprint((<BUS_ID_BAR>, a, b)) == fingerprint((0, <BUS_ID_BAR>, a,
b))`, so if `BUS_ID_FOO` is `0`, the receiver could be ambiguous.

This is fixed now, because every bus interaction will at least
fingerprint the bus ID, which is always the first element.