Separate Bus Receives from Identities (#2402)
#2361 introduced a new `Identity` type, which could be - among other
things - a bus send or a bus receive. Also, there was another map from
bus ID to receive.

Actually, the receives are not needed in the list of identities, so I
removed them from the type entirely. They were ignored everywhere
anyway. As a bonus, we can now also remove the
`BusReceive::interaction_id`.