Add accumulator columns to `PhantomBusInteraction` (#2412)
In our manual bus witgen, we currently assume that all second-stage
witness columns in source order are always of the form: "folded"
payloads for bus interaction 1, accumulators for bus interaction 1,
"folded" payloads for bus interaction 2, ...

As we implement #2337, this will no longer be the case (we might not
have a folded payload, or there might be two interactions using the same
accumulator).

I tried to write code to find the accumulators from the polynomial
identities, but I failed. So now, they are simply part of the
`PhantomBusInteraction` annotation, which I think is fair. I still hope
that we can automatically find the "folded" columns if they exist.