Add `Fp` to extension field implementation (#2427)
In our PIL std lib, we have an `Ext` that abstracts over `Fp2` and
`Fp4`. But in cases where we don't need any extension field (e.g.
BN254), we wrapped base field elements as `Fp2` anyway. This is a bit
weird and can lead to issues when the optimizer removes the unused
dimension (e.g., it might still be referenced by a prover function or
witgen annotation). I had this problem in #2426.

Now, there is an `Ext<T>::Fp` variant that simply wraps `T`.