Compute multiple cell values conditionally from other cell values. (#2434)
I think this is the last part for prover functions, which is needed for
the arithmetic machine:

Compute the values of multiple cells at the same time from other known
cells, but only if a condition is true. The biggest challenge here is
that we need to be able to evaluate the condition at compile time.