Bus single interaction update (#2459)
Depends on #2458, which depends on #2457.

Instead of removing the single bus interaction API entirely, I kept it
as a user friendly choice (so that the user can just provide raw inputs
in the case of a single bus interaction, instead of providing a vector
of tuple of inputs). This PR makes the single version call the multi
version (and falls to the "odd" case) to ensure maximum code reuse and
best optimization.