Bus multi interaction linker (#2468)
Ready for review. Depends on #2459, which depends on #2457.

1. Added `lookup_multi_receive` and `permutation_multi_receive` APIs for
the linker and test files to use.
2. Modified linker to use the multi version by default. This in practice
reduces the number of bus interactions to 2 or 3 (depending if we only
have permutation, only have lookup, or have both). This can be furthur
optimized by directly using `bus_multi_receive` function in the linker,
but requires more processing of lower level expressions, which I'd leave
to another PR. In practice, this would reduce the number of bus
interactions from 2 or 3 to always 2. This is currently being
implemented in #2469.
3. Added test cases to the pipeline.