Stwo publics (#2438)
# Extend `stwo` to Support Public Input

## Description
This PR extends `stwo` to support public input.

While `stwo` does not natively implement the public input feature, it is
feasible to achieve this by implementing it in the `stwo` backend.

## Implementation Details
The public values are formed using a selector to pick specific values
from a witness column, which are then made public to the verifier.

This logic is already implemented in the `Plonky3` backend, where public
values are passed directly as inputs to the `prove` and `verify`
functions, simplifying the process.

## ✅ To-Do List
To achieve similar functionality in `stwo`, the following steps need to
be implemented:

- [x] Add selector columns in setup.
- [x] Add constraints linking the public columns, the selector, and the
corresponding witness columns.
- [x] pass public values to verification.

---------

Co-authored-by: Georg Wiese <georgwiese@gmail.com>
Co-authored-by: Thibaut Schaeffer <schaeffer.thibaut@gmail.com>