Bus interaction simplify API (#2483)
Ready for review. Depends on #2479 merged, which is currently in a chain
of 3 PRs that need merging.

Inspired by Chris' comment here:
https://github.com/powdr-labs/powdr/pull/2465#discussion_r1954166247

This PR will remove anything in the like of
`bus_send(BusInteraction::Send(...))`, because the "send" is duplicated
information. Instead, it will just be `bus` or `bus_multi` that handles
both the send and the receive cases.

Linker and tests that depend on these APIs are also updated.