Do not skip the current queue item. (#2496)
The idea behind the "skip" feature was that if we have an update to a
variable coming from solving an identity, this variable causes all
identities that contain the variable to be re-added to the queue. The
identity we currently processed is among those. I thought we do not need
to re-add it because we just processed it, but it turns out that in some
situations, you can derive more information from the same queue item,
for example when a derived range constraint on a variable together with
existing constraints allows you to fully solve the identity.