Add type parameter to `MachineCallVariable` (#2527)
Extracted from #2488, to keep the diff smaller

In #2488, I want to identify a `MachineCallVariable` by a `bus_id: T`
instead of an `identity_id: u64`. As a result, it gets a new `<T>`
parameter, which propagates to `Variable` and quite far into the
codebase. This PR adds it already. The changes here should be quite
mechanic.