Bug fix: Include intermediates from receives in machine parts (#2546)
Extracted from #2541

Fixes a bug that the machine parts doesn't include the intermediate
definitions that only appear in one of the receives. An example of this
is the [Arith
large](https://github.com/powdr-labs/powdr/blob/main/std/machines/large_field/arith.asm)
machine, which has intermediates like `x1c` in their list of arguments,
but doesn't refer to the in any constraint.

---------

Co-authored-by: chriseth <chris@ethereum.org>