`Keccakf32Memory`: use `compute_from` instead of `provide_value` (#2553)
This PR changes the usage of `std::prover::provide_value` in
`Keccakf32Memory` to `std::prover::compute_from`, in the hope of being
able to use the JIT for this machine in #2541.

~Witgen time goes of `test_data/std/keccakf32_memory_test.asm` from 60s
to 76s, probably because too many things are being evaluated.~
This is fixed by 1c1564783e3ad0f90a3bf139ff90120befc0c14f.