Implement dynamic bus (JIT) (#2545)
With this PR, there is no more `.unwrap()` on `BusSend::static_bus_id()`
(previously `bus_id()`) and `BusSend::try_match_static()`.

Note that this will only generate a machine call when the bus ID is
"known concrete". Getting there might require branching. To relax this
(and require only known, but not concrete), I think we could change
[`Effect::MachineCall`](https://github.com/powdr-labs/powdr/blob/main/executor/src/witgen/jit/effect.rs#L25)
to take a symbolic expression instead of a concrete value, but
[`CanProcessCall::can_process_call_fully()`](https://github.com/powdr-labs/powdr/blob/d039c7bd819c59fd5b44dd6e971a65f5ae4780e7/executor/src/witgen/jit/witgen_inference.rs#L625)
still expects a concrete value, which can't easily be changed, so I
think this is the best we can do.