Create bus.asm call to expose native BusInteraction to the backend (#2510)
Currently just the bus.asm call is created. Next step will be editing
`bus.rs` linker to do this call for backends that needs the native bus
interaction.

A key quesiton remains whether we need to import the backend crate or
create another linker mode in the linker crate in order to distinguish
which cases to pass down the native bus interaction vs which cases to
call `bus_multi_linker`.

A third solution is to make both `bus_multi_linker` and
`bus_native_linker` call in `bus.rs`, and this will create both
`BusInteraction` and `PhantomBusInteraction` for all backends, and some
backends can just ignore the native `BusInteraction`, similar to what
we've done for native lookup and permutation.