Powdr PIL runs witgen with mock backend by default (#2601)
Now that witgen depends on backend type, `powdr pil` requires a backend
type to run through witgen. It only runs till optimized pil if
`--prove-with` is not provided.

To match the description, this PR defaults backend to Mock if
`--prove-with` is not provided, so `powdr pil` can still run till
witgen. To prove with any backend, `--prove-with` is still required.